refactor: исправить ошибки автотеста 13-sprint

+++ controllers/cards.js
@@ -12,19 +12,22 @@ module.exports.getCards = (req, res) => {
 };
 
 // удалить карточку
+// eslint-disable-next-line consistent-return
 module.exports.deleteCard = (req, res) => {
   const { cardId } = req.params;
+
+  if (!cardId) {
+    return res.status(400).json({ message: 'Некорректный id карточки' });
+  }
 
   Card.findByIdAndRemove(cardId)
     .then((deletedCard) => {
       if (!deletedCard) {
-        return res.status(404).send({ message: 'Карточка не найдена' });
+        return res.status(404).json({ message: 'Карточка не найдена' });
       }
-      return res.status(200).send({ message: 'Карточка успешно удалена' });
+      return res.status(200).json({ message: 'Карточка успешно удалена' });
     })
-    .catch((err) => {
-      res.status(500).send({ message: err });
-    });
+    .catch((err) => res.status(500).json({ message: err }));
 };
 
 // создать карточку
@@ -43,9 +46,15 @@ module.exports.postCard = (req, res) => {
 };
 
 // лайкнуть карточку
+// eslint-disable-next-line consistent-return
 module.exports.likeCard = (req, res) => {
   const { cardId } = req.params;
   const { userId } = req.user;
+
+  if (!cardId) {
+    return res.status(400).json({ message: 'Некорректный id карточки' });
+  }
+
   Card.findByIdAndUpdate(
     cardId,
     { $addToSet: { likes: userId } },
@@ -53,19 +62,23 @@ module.exports.likeCard = (req, res) => {
   )
     .then((updatedCard) => {
       if (!updatedCard) {
-        return res.status(404).send({ message: 'Карточка не найдена' });
+        return res.status(404).json({ message: 'Карточка не найдена' });
       }
-      return res.status(200).send({ data: updatedCard });
+      return res.status(200).json({ data: updatedCard });
     })
-    .catch((err) => {
-      res.status(400).json({ message: err });
-    });
+    .catch((err) => res.status(500).json({ message: err }));
 };
 
 // убрать лайк карточки
+// eslint-disable-next-line consistent-return
 module.exports.dislikeCard = (req, res) => {
   const { cardId } = req.params;
   const { userId } = req.user;
+
+  if (!cardId) {
+    return res.status(400).json({ message: 'Некорректный id карточки' });
+  }
+
   Card.findByIdAndUpdate(
     cardId,
     { $pull: { likes: userId } },
@@ -73,11 +86,11 @@ module.exports.dislikeCard = (req, res) => {
   )
     .then((updatedCard) => {
       if (!updatedCard) {
-        return res.status(404).send({ message: 'Карточка не найдена' });
+        return res.status(404).json({ message: 'Карточка не найдена' });
       }
-      return res.status(200).send({ data: updatedCard });
+      return res.status(200).json(updatedCard);
     })
     .catch((err) => {
-      res.status(400).send({ message: err });
+      res.status(500).json({ message: err.message });
     });
 };
